Improved albedo method and snow outputs
1. The albedo improvements discussed in the last commit are expanded to be more independent of the model timestep.

2. The choice between accumulation-season or melt-season decay curves is now determined by whether the snow surface has EVER reached 0 deg. C since the last snowfall. Previously, the albedo was temperature-dependent and could fluctuate on a diurnal cycle, with higher albedo at night because the lower temperature caused the model to switch to the "accumulation-season" decay curve. This was unphysical, and now snowfall is the ONLY thing that can cause the albedo to increase.

3. The Aggregated.Values file has been improved to display various useful snow statistics. The HasSnow column now reports the fraction of the watershed with snow cover, and most snow variables are normalized only across the snow-covered area (i.e., the average snowpack temperature was previously averaged across the whole area, even non-snow-covered cells, but now it's only averaged across snow-covered cells). Mass-related variables (SWE, vapor flux, melt) are still normalized over the whole area.

+++ DHSVM/sourcecode/StoreModelState.c
@@ -270,7 +270,7 @@ void StoreModelState(char *Path, DATE * Current, MAPSIZE * Map,
   for (y = 0; y < Map->NY; y++) {
     for (x = 0; x < Map->NX; x++) {
       if (INBASIN(TopoMap[y][x].Mask))
-        ((float *)Array)[y * Map->NX + x] = (float)SnowMap[y][x].LastSnow;
+        ((float *)Array)[y * Map->NX + x] = SnowMap[y][x].LastSnow;
       else
         ((float *)Array)[y * Map->NX + x] = NA;
     }
